split message and frame parsing
