+++ README.md
@@ -25,9 +25,8 @@ const url = new URL("/lenkan/cesr-js/refs/heads/main/fixtures/geda.cesr", "https
 const response = await fetch(url);
 
 if (response.body) {
-  for await (const message of parse(response.body)) {
-    console.log(message.payload);
-    console.log(message.attachments);
+  for await (const frame of parse(response.body)) {
+    console.log(frame);
   }
 }
 ```
@@ -51,5 +50,4 @@ curl https://raw.githubusercontent.com/lenkan/cesr-js/refs/heads/main/fixtures/g
 
 - [x] Simple parsing of KERI CESR Stream
 - [ ] API for encoding/decoding primitives
-- [ ] Cryptographic verification of CESR Stream
 - [ ] ...
